Update neovim/init.vim
Changed the autocomplete interface so that when <TAB> is pressed it does
to the first, the next item in the autocomplete list, <S-TAB> or
Shift+TAB will go to the previous item in the autocomplete list.
Additionally, to select the highlighted item for autocomplete, <CR> or
Enter can be pressed to apply the recommended autocomplete item.